+++ bin/Execute_report.R
@@ -26,11 +26,13 @@ option_list = list(
     make_option(c("-z", "--rlog"), action="store_true", default=TRUE, help="Use rlog instead of vst normalization."),
     make_option(c("-n", "--nsub_genes"), type="integer", default=NULL, help="Subset number of genes for vst."),
 
-    make_option(c("-a", "--pathway_analysis"), action="store_true", default=FALSE, help="Pathway analysis."),
+    make_option(c("-a", "--pathway_analysis"), action="store_true", default=FALSE, help="Whether to run pathway analysis."),
     make_option(c("-g", "--organism"), type="character", default=NULL, help="Organism, e.g. Hsapiens."),
     make_option(c("-i", "--species_library"), type="character", default=NULL, help="Library name. Example format: org.At.tair.db", metavar="character"),
     make_option(c("-u", "--keytype"), type="character", default=NULL, help="Keytype. Example format: TAIR (varies greatly depending on library!)", metavar="character"),
     make_option("--custom_gmt", type="character", default=NULL, help="Path to custom GMT file to query during pathway analysis.", metavar="character"),
+    make_option("--set_background", action="store_true", default=TRUE, help="Whether to use a background list for pathway analysis; if true, will only consider expressed genes (i.e. mean counts > 0) for PA."),
+    make_option("--custom_background", type="character", default=NULL, help="Path to a custom background list TXT for pathway analysis; if provided, will only consider genes in that list for PA."),
     make_option(c("-w", "--min_DEG_pathway"), type="integer", default=NULL, help="min. number of genes DE in a pathway for this pathway to be considered enriched.", metavar="integer"),
 
     make_option(c("-s", "--proj_summary"), type="character", default=NULL, help="Project summary file", metavar="character"),
@@ -76,6 +78,8 @@ rmarkdown::render(opt$report, output_file = opt$output, knit_root_dir = wd, outp
                                 organism = opt$organism,
                                 species_library = opt$species_library,
                                 custom_gmt = opt$custom_gmt,
+                                set_background = opt$set_background,
+                                custom_background = opt$custom_background,
                                 keytype = opt$keytype,
                                 min_DEG_pathway = opt$min_DEG_pathway,